Changed decimal fields to float in hall model since it was truncating data.

+++ hall/models.py
@@ -10,14 +10,14 @@ class hall(models.Model):
     date = models.DateTimeField()
     temperature = models.DecimalField(max_digits=7, decimal_places=2, blank=True, default=300.0)
     symmetry_factor = models.DecimalField(max_digits=7, decimal_places=2, blank=True, default=1.0)
-    sheet_coefficient = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
-    sheet_resistance = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
-    sheet_concentration = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
+    sheet_coefficient = models.FloatField(blank=True, null=True)
+    sheet_resistance = models.FloatField(blank=True, null=True)
+    sheet_concentration = models.FloatField(blank=True, null=True)
     thickness = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
-    mobility = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
-    bulk_coefficient = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
-    bulk_resistance = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
-    bulk_concentration = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
+    mobility = models.FloatField(blank=True, null=True)
+    bulk_coefficient = models.FloatField(blank=True, null=True)
+    bulk_resistance = models.FloatField(blank=True, null=True)
+    bulk_concentration = models.FloatField(blank=True, null=True)
 
     def __unicode__(self):
         return self.date
